backend/pyproject.toml
[project]
name = "notification-creator"
version = "0.1.0"
description = "Internal notification service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx", "pytest-asyncio"]
